+++ packages/cater-emotion/plugin.js
@@ -1,0 +1,10 @@
+// const cater = require('cater');
+
+module.exports = function(cater, options) {
+
+  cater.on('configured', () => {
+    console.log("configured");
+  });
+
+  return true;
+}