+++ packages/cater-jest/index.js
@@ -1,4 +1,4 @@
-const clone = require('clone');
+const clone = require("clone");
 const path = require("path");
 
 const assetTransformer = path.join(__dirname, "asset-transformer.js");
@@ -6,15 +6,13 @@ const emptyModule = path.join(__dirname, "empty-module.js");
 const resolver = path.join(__dirname, "resolver.js");
 const transformer = path.join(__dirname, "transformer.js");
 
-const assetExtensions = ['jpg', 'jpeg', 'gif', 'png', 'svg'];
-const assetMatch = `\\.(${assetExtensions.join('|')})$`;
+const assetExtensions = ["jpg", "jpeg", "gif", "png", "svg"];
+const assetMatch = `\\.(${assetExtensions.join("|")})$`;
 
 const config = {
   moduleNameMapper: {},
   resolver: resolver,
-  transform: {
-    "\\.js$": transformer,
-  },
+  transform: { "\\.js$": transformer },
   transformIgnorePatterns: ["node_modules/(?!(cater)/)"]
 };
 
@@ -23,4 +21,4 @@ config.moduleNameMapper[assetMatch] = emptyModule;
 
 module.exports = function() {
   return clone(config);
-}
+};