+++ packages/cater/src/commands.js
@@ -7,6 +7,7 @@ import webpackBuild from "./webpack-build";
 // Returns a promise to a production build of the application.
 module.exports.runBuild = function() {
   this.build = true;
+  this.production = true;
 
   return fs
     .remove(this.buildPath)